remove old intro banner, add logic to combine discount percentages
