--- client/components/jetpack/intro-pricing-banner-v2/style.scss
@@ -1,59 +1,0 @@
-.intro-pricing-banner-v2,
-.intro-pricing-banner-v2__sticky {
-	display: flex;
-	flex-direction: row;
-
-	justify-content: center;
-
-	.intro-pricing-banner-v2__discount,
-	.intro-pricing-banner-v2__guarantee {
-		display: flex;
-		flex-direction: row;
-		align-items: center;
-
-		font-size: $font-body;
-	}
-
-	> :not( :last-child ) {
-		margin-right: 16px;
-	}
-}
-
-.intro-pricing-banner-v2 {
-	height: 54px;
-	margin-bottom: 32px;
-}
-
-.intro-pricing-banner-v2__sticky {
-	position: fixed;
-	width: 100%;
-	left: 0;
-	top: var( --masterbar-height );
-
-	background-color: var( --color-surface );
-	box-shadow: 0 4px 4px 0 #00000014;
-
-	z-index: 2;
-	height: 54px;
-}
-.is-section-jetpack-connect,
-.is-group-jetpack-cloud.is-section-jetpack-cloud-pricing {
-	.intro-pricing-banner-v2__sticky {
-		top: 0;
-	}
-}
-
-// target Calypso pricing filter bar, not jetpack cloud
-.is-group-sites.is-section-plans .intro-pricing-banner-v2__sticky {
-	// At this screen size, the left sidebar is 228px;
-	@include breakpoint-deprecated( '>660px' ) {
-		width: calc( 100% - 228px );
-		left: 228px;
-	}
-
-	// At this screen size, the lft sidebar is 272px;
-	@include breakpoint-deprecated( '>960px' ) {
-		width: calc( 100% - 272px );
-		left: 272px;
-	}
-}
